requirements.txt
alembic==1.3.2
atomicwrites==1.3.0
attrs==19.3.0
awesome-slugify==1.6.5
bcrypt==3.1.7
bleach==3.1.0
certifi==2019.11.28
cffi==1.13.2
Click==7.0
cloudinary==1.19.1
colorama==0.4.3
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Markdown==0.3
Flask-Migrate==2.5.2
Flask-MySQL==1.4.0
Flask-SimpleMDE==0.3.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
importlib-metadata==1.3.0
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mock==3.0.5
more-itertools==8.0.2
mysql-connector-python==8.0.18
packaging==19.2
Pillow==6.2.1
pluggy==0.13.1
protobuf==3.11.2
py==1.8.0
pycparser==2.19
PyMySQL==0.9.3
pyparsing==2.4.5
pytest==5.3.2
python-dateutil==2.8.1
python-editor==1.0.4
regex==2019.12.20
six==1.13.0
SQLAlchemy==1.3.12
Unidecode==1.1.1
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
wtf-tinymce==0.1.3
WTForms==2.2.1
zipp==0.6.0
